remove original file ops
