+++ src/utils.ts
@@ -10,110 +10,6 @@
 */
 
 /**
- * Indexed DB configurations
- */
-export const DB_NAME = 'pv_db';
-export const STORE_NAME = 'pv_store';
-export const PV_FILE_STORE = 'pv_file';
-export const DB_VERSION = 2;
-
-/**
- * Storage Interface.
- */
-interface PvStorage {
- setItem: (key: string, value: any) => void | Promise<void>;
- getItem: (key: string) => any | Promise<any>;
- removeItem: (key: string) => void | Promise<void>;
- exists: (key: string) => boolean | Promise<boolean>;
-}
-
-/**
- * Opens indexedDB connection, handles version changes and gets the db instance.
- *
- * @returns The instance of indexedDB connection.
- */
-export function getDB(): Promise<IDBDatabase> {
-  return new Promise((resolve, reject) => {
-    const request = self.indexedDB.open(DB_NAME, DB_VERSION);
-    request.onerror = () => {
-      reject(request.error);
-    };
-    request.onsuccess = () => {
-      resolve(request.result);
-    };
-    request.onupgradeneeded = () => {
-      if (!request.result.objectStoreNames.contains(STORE_NAME)) {
-        request.result.createObjectStore(STORE_NAME);
-      }
-      if (!request.result.objectStoreNames.contains(PV_FILE_STORE)) {
-        request.result.createObjectStore(PV_FILE_STORE);
-      }
-    };
-  });
-}
-
-export const requestHelper = (request: IDBRequest): Promise<any> => new Promise((resolve, reject) => {
-  request.onerror = () => {
-    reject(request.error);
-  };
-  request.onsuccess = () => {
-    resolve(request.result);
-  };
-});
-
-/**
- * Gets the storage to use. Either tries to use IndexedDB or localStorage.
- * @param requireDB Flag indication if IndexedDB is required.
- *
- * @returns PvStorage instance to use as storage.
- */
-export function getPvStorage(requireDB = false): PvStorage {
-  if (self.indexedDB) {
-    return {
-      setItem: async (key: string, value: any) => {
-        const db = await getDB();
-        const request = db.transaction(STORE_NAME, 'readwrite').objectStore(STORE_NAME).put(value, key);
-        await requestHelper(request);
-        db.close();
-      },
-      getItem: async (key: string) => {
-        const db = await getDB();
-        const request = db.transaction(STORE_NAME, 'readonly').objectStore(STORE_NAME).get(key);
-        const res = await requestHelper(request);
-        db.close();
-        return res;
-      },
-      removeItem: async (key: string) => {
-        const db = await getDB();
-        const request = db.transaction(STORE_NAME, 'readwrite').objectStore(STORE_NAME).delete(key);
-        await requestHelper(request);
-        db.close();
-      },
-      exists: async (key: string) => {
-        const db = await getDB();
-        const request = db.transaction(STORE_NAME, 'readonly').objectStore(STORE_NAME).count(key);
-        const res = await requestHelper(request);
-        db.close();
-        return res > 0;
-      }
-    };
-  } else if (requireDB) {
-    throw new Error("IndexedDB is required but not available.");
-  } else if (self.localStorage) {
-    const { setItem, getItem, removeItem } = self.localStorage;
-    const exists = (key: string) => getItem(key) !== null;
-    return {
-      setItem,
-      getItem,
-      removeItem,
-      exists
-    };
-  }
-
-  throw new Error("Cannot get a persistent storage object.");
-}
-
-/**
  * Convert a null terminated phrase stored inside an array buffer to a string
  *
  * @param arrayBuffer input array buffer
